LayoutTests/fast/viewport/ios/width-is-device-width-overflowing-no-shrink-to-fit.html

<!DOCTYPE html>

<html>
<head>
    <meta name="viewport" content="width=device-width, shrink-to-fit=no">
    <script src="resources/viewport-test-utils.js"></script>
    <style>
        .wide {
            height: 1px;
            width: 1000px;
            background-color: gray;
        }
    </style>
</head>
<body onload="runTest()">

<p>Viewport: <span id="viewport"></span></p>
<div id="result"></div>

<div class="wide"></div>

</body>
</html>
