LayoutTests/fast/viewport/ios/width-is-device-width-overflowing-no-shrink-to-fit-expected.txt

Viewport: width=device-width, shrink-to-fit=no

scale	1.00000
maxScale	5.00000
minScale	1.00000
visibleRect	{"left":"0.00000","top":"0.00000","width":"320.00000","height":"480.00000"}
